pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agents"
version = "0.1.0"
dependencies = [
    "typing_extensions>=4.10",
    "openai==1.82.0",
    "tiktoken==0.9.0",
    "python-dotenv==1.1.0",
    "langchain==0.3.25",
    "langchain-openai==0.3.18",
    "langgraph==0.4.7"
]

[project.optional-dependencies]
mcp = [
    "fastapi",
    "uvicorn",
    "langchain-mcp-adapters==0.0.11",
]
studio = [
    "langgraph-cli[inmem]==0.3.3",
    "langsmith==0.3.42"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
